Move helper function to new utils/template module
