+++ src/distilabel/steps/tasks/text_generation.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from jinja2 import Template
@@ -21,6 +20,7 @@ from pydantic import Field, PrivateAttr
 from distilabel.errors import DistilabelUserError
 from distilabel.steps.tasks.base import Task
 from distilabel.utils.chat import is_openai_format
+from distilabel.utils.template import check_column_in_template
 
 if TYPE_CHECKING:
     from distilabel.steps.tasks.typing import ChatType
@@ -377,33 +377,3 @@ class ChatGeneration(Task):
         """The output is formatted as a dictionary with the `generation`. The `model_name`
         will be automatically included within the `process` method of `Task`."""
         return {"generation": output}
-
-
-def check_column_in_template(
-    column: str, template: str, page: str = "components-gallery/tasks/textgeneration/"
-) -> None:
-    """Checks if a column is present in the template, and raises an error if it isn't.
-
-    Args:
-        column: The column name to check in the template.
-        template: The template of the Task to be checked, the input from the user.
-        page: The page to redirect the user for help . Defaults to "components-gallery/tasks/textgeneration/".
-
-    Raises:
-        DistilabelUserError: Custom error if the column is not present in the template.
-    """
-    pattern = (
-        r"(?:{%.*?\b"
-        + re.escape(column)
-        + r"\b.*?%}|{{\s*"
-        + re.escape(column)
-        + r"\s*}})"
-    )
-    if not re.search(pattern, template):
-        raise DistilabelUserError(
-            (
-                f"You required column name '{column}', but is not present in the template, "
-                "ensure the 'columns' match with the 'template' to avoid errors."
-            ),
-            page=page,
-        )
